requirements.txt
vncorenlp
transformers
sentencepiece
emoji==2.14.1
Flask==3.1.1
flask_cors==6.0.0
Flask-Limiter==3.8.0
matplotlib
numpy
pandas
regex==2024.11.6
requests>=2.32.3
chardet<6
urllib3<3
torch
underthesea
wordcloud==1.9.4
protobuf==3.20.3
pydantic==2.11.3
pytest>=8.0.0
prometheus_client>=0.20.0
gunicorn>=22.0.0
gTTS>=2.5.4
onnxruntime>=1,<2
piper-tts>=1.3.0
